+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - The `uuid.UUID`-type has been added to `IdType`s.
+- Add the `Translator.enable_uuid_heuristics` flag (default=`False`).
 
 ### Fixed
 - Duplicate explicit names are now supported for most types (closes #4).